Update tests to "work"
Something very wrong happens when you work with unconstrained and
constrainted problems in the same execution, even when you call
terminate. Possibly something left in the memory of CUTEst breaks
the next problem.
If you run a constrained problem first, the next unconstrained
problem does not work well, and vice-versa. It runs, but the
objective values and gradient, and constraint values are NaN, or
zero, or garbage.